+++ ADPTest.SharedKernel/Dtos/TaskDto.cs
@@ -10,7 +10,7 @@ namespace ADPTest.SharedKernel.Dtos
     {
         public string Id { get; set; }
         public string Operation { get; set; }
-        public long Left { get; set; }
-        public long Right { get; set; }
+        public double Left { get; set; }
+        public double Right { get; set; }
     }
 }
